Solved merge conflicts
