+++ src/main/java/cpp/projects/ProjectList.java
@@ -54,4 +54,45 @@ public class ProjectList {
         }
         System.out.println("____________________________________________________________");
     }
+
+    /**
+     * Add a todo to one project in projectList.
+     * @param indexString Index of the project
+     */
+
+    public void addTodoToProject(String indexString, String todoString) {
+        int index;
+        try {
+            index = Integer.parseInt(indexString);
+        } catch (NumberFormatException e) {
+            System.out.println("The project index is not an integer!");
+            return;
+        }
+
+        Project targetProject = new Project();
+        targetProject = projectList.get(index - 1);
+
+        targetProject.addTodo(todoString);
+        System.out.println("Todos have been added to project " + targetProject.getTitle());
+    }
+
+    /**
+     * Marks a todo in a project as done.
+     * @param indexStringProj Index of the project.
+     * @param indexStringTodo Index of the todo.
+     */
+
+    public void markTodoAsDone(String indexStringProj, String indexStringTodo) {
+        int indexProj;
+        int indexTodo;
+        try {
+            indexProj = Integer.parseInt(indexStringProj);
+            indexTodo = Integer.parseInt(indexStringTodo);
+        } catch (NumberFormatException e) {
+            System.out.println("The input is not an integer!");
+            return;
+        }
+        Project targetProject = projectList.get(indexProj - 1);
+        targetProject.markTodoAsDone(indexTodo);
+    }
 }
