+++ src/main/java/cpp/projects/commandhandler/CommandHandler.java
@@ -1,6 +1,8 @@
 package cpp.projects.commandhandler;
 
+import cpp.exceptions.IllegalCommandException;
 import cpp.projects.ProjectList;
+import cpp.Messages;
 
 import java.util.Scanner;
 
@@ -26,7 +28,7 @@ public class CommandHandler {
      *
      * @param projectList ProjectList for commands to work with
      */
-    public void handleUserInput(ProjectList projectList, String userInput) {
+    public void handleUserInput(ProjectList projectList, String userInput) throws IllegalCommandException {
         commands = userInput.split(" ");
 
         switch (commands[0].toLowerCase()) {
@@ -45,6 +47,19 @@ public class CommandHandler {
         case "viewproject": //view all project(s) by name
             projectList.printProject();
             break;
+        case "todo":
+            if (commands.length < 3) {
+                throw new IllegalCommandException(Messages.MESSAGE_INVALID_COMMAND_FORMAT);
+            }
+            String todoString = parseTodoString(commands);
+            projectList.addTodoToProject(commands[1], todoString);
+            break;
+        case "mark":
+            if (commands.length < 3) {
+                throw new IllegalCommandException(Messages.MESSAGE_INVALID_COMMAND_FORMAT);
+            }
+            projectList.markTodoAsDone(commands[1], commands[2]);
+            break;
         default:
             System.out.println("____________________________________________________________");
             System.out.println("Unknown command.");
@@ -52,5 +67,17 @@ public class CommandHandler {
             break;
         }
 
+    }
+
+    public String parseTodoString(String[] strings) {
+        String todoString = "";
+        if (strings.length == 3) {
+            todoString = strings[2];
+        } else {
+            for (int i = 2; i < strings.length; i++) {
+                todoString = todoString + " " + strings[i];
+            }
+        }
+        return todoString;
     }
 }
